requirements.txt
discord.py>=2.3.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
httpx>=0.24.0
openai>=1.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
authlib>=1.3.0
itsdangerous>=2.1.0
jinja2>=3.1.0
python-multipart>=0.0.6